// src/types/transaction.type.ts
import type { TransactionType } from "@prisma/client";
import type { CategorySummary } from "./category.types";

export interface TransactionFilter {
	userId: string;
	date?: {
		gte: Date;
		lte: Date;
	};
	type?: TransactionType;
	categoryId?: string;
}

export interface TransactionSummary {
	totalExpense: number;
	totalIncome: number;
	balance: number;
	expenseByCategory: CategorySummary[];
}